refactor: moved search folder and created components and utils for them
