+++ packages/views/screens/router/routesMap.js
@@ -26,7 +26,7 @@ import EditVote from '@screens/managers/editVote';
 import VotingQueue from '@screens/managers/votingQueue';
 import DeviceDisconnect from '@screens/managers/deviceDisconnectDialog';
 import NewReleaseDialog from '@updater/detail/info/newReleaseDialog';
-import SearchBar from '@shared/searchBar';
+import SearchBar from '@search/components/SearchBar';
 import ReclaimBalance from '@legacy/manager/reclaimBalance';
 import ReclaimBalanceModal from '@legacy/manager/reclaimBalanceModal';
 import MultiSignature from '@screens/managers/multiSignature';
