+++ setup/i18n/locales/en/common.json
@@ -3,7 +3,6 @@
   " is out. ": " is out. ",
   "\"{{username}}\" is already taken.": "\"{{username}}\" is already taken.",
   "64 bytes left": "64 bytes left",
-  "A bit more. Make sure to type at least 3 characters.": "A bit more. Make sure to type at least 3 characters.",
   "A great way to start is to top up your account with some {{value}}.": "A great way to start is to top up your account with some {{value}}.",
   "About": "About",
   "Access your account by scanning the QR code below with the Lisk Mobile App:": "Access your account by scanning the QR code below with the Lisk Mobile App:",
@@ -57,7 +56,6 @@
   "Based on your available balance and rounded down to a multiple of 10 LSK, your total remaining balance is {{maxAmount}} LSK": "Based on your available balance and rounded down to a multiple of 10 LSK, your total remaining balance is {{maxAmount}} LSK",
   "Before you continue using Lisk, please read and accept the": "Before you continue using Lisk, please read and accept the",
   "Below are the details of your locked balances and the unlock waiting periods. From here you can submit an unlock transaction when waiting periods are over.": "Below are the details of your locked balances and the unlock waiting periods. From here you can submit an unlock transaction when waiting periods are over.",
-  "Block": "Block",
   "Block ID": "Block ID",
   "Block height": "Block height",
   "Blocks": "Blocks",
@@ -275,7 +273,6 @@
   "No {{activeTab}} delegates found.": "No {{activeTab}} delegates found.",
   "Nonce": "Nonce",
   "Nothing available to unlock": "Nothing available to unlock",
-  "Nothing has been found. Make sure to double check the ID you typed.": "Nothing has been found. Make sure to double check the ID you typed.",
   "Now you can send it to the blockchain. You may also copy or download it, if you wish to send the transaction using another device later.": "Now you can send it to the blockchain. You may also copy or download it, if you wish to send the transaction using another device later.",
   "Number": "Number",
   "Number of signatures must be above {{num}}.": "Number of signatures must be above {{num}}.",
@@ -377,8 +374,6 @@
   "Save it on an encrypted hard drive: USB key or a backup drive": "Save it on an encrypted hard drive: USB key or a backup drive",
   "Save your passphrase": "Save your passphrase",
   "Scan address": "Scan address",
-  "Search": "Search",
-  "Search within the network...": "Search within the network...",
   "Security": "Security",
   "Select All": "Select All",
   "Select device": "Select device",
