+++ src/modules/search/components/BlockResultList/BlockResultList.css
