+++ src/modules/search/components/BlockResultList/index.js
@@ -1,0 +1,1 @@
+export { default } from './BlockResultList';
