+++ src/modules/search/components/Search/index.js
@@ -1,0 +1,1 @@
+export { default } from './Search';
