+++ src/modules/search/components/SearchBar/SearchBar.css
@@ -1,4 +1,4 @@
-@import '../../../../setup/react/app/mixins.css';
+@import '../../../../../setup/react/app/mixins.css';
 
 .wrapper {
   width: calc(100% - 456px);
