+++ src/modules/search/components/SearchBar/index.js
@@ -1,0 +1,1 @@
+export { default } from '../../manager/searchBarManager';
