+++ src/modules/search/components/TransactionResultList/TransactionResultList.css
@@ -1,0 +1,90 @@
+.wrapper {
+  width: 100%;
+  height: auto;
+  box-sizing: border-box;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: flex-start;
+  padding: 30px 0 var(--horizontal-padding-s);
+  background: var(--color-white);
+  border-radius: 0 0 var(--border-radius-box) var(--border-radius-box);
+  border-top: solid 1px var(--color-mystic);
+  margin-top: -10px;
+}
+
+.header {
+  width: 100%;
+  height: 100%;
+  box-sizing: border-box;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: flex-start;
+  margin-bottom: 8px;
+  padding: 2px 20px 12px;
+
+  & label {
+    @mixin contentLargest bold;
+
+    color: var(--color-maastricht-blue);
+  }
+}
+
+.subTitles {
+  width: 100%;
+  height: auto;
+  box-sizing: border-box;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.content {
+  @mixin contentLargest;
+
+  width: 100%;
+  height: auto;
+  max-height: 220px;
+  box-sizing: border-box;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: flex-start;
+  margin: 0;
+
+  & .resultRow > img {
+    margin: 10px 20px 10px 0;
+  }
+
+  & .transactionId {
+    color: var(--color-maastricht-blue);
+  }
+
+  & .transactionMessage {
+    @mixin contentLargest bold;
+
+    color: var(--color-ink-blue);
+    box-sizing: border-box;
+    margin-left: auto;
+  }
+
+  & .resultRow {
+    width: 100%;
+    height: auto;
+    min-height: 53px;
+    padding: 5px 20px;
+    box-sizing: border-box;
+    display: flex;
+    flex-direction: row;
+    justify-content: flex-start;
+    align-items: center;
+    cursor: pointer;
+    color: var(--color-maastricht-blue);
+
+    &.active,
+    &:hover {
+      background-color: var(--color-rows-hover);
+    }
+  }
+}
