+++ src/modules/search/components/TransactionResultList/TransactionResultList.js
@@ -1,19 +1,9 @@
 import React from 'react';
-import { MODULE_ASSETS_MAP } from '@transaction/configuration/moduleAssets';
 import Icon from '@basics/icon';
 import { truncateAddress } from '@wallet/utilities/account';
 import LiskAmount from '@shared/liskAmount';
-import styles from './transactionsAndBlocks.css';
-
-const getTxConfig = (t, transactions) => {
-  const { asset, fee } = transactions[0];
-
-  return {
-    icon: transactions[0].moduleAssetId === '2:0' ? undefined : MODULE_ASSETS_MAP['2:0'].icon,
-    subTitle: transactions[0].moduleAssetId === '2:0' ? t('Amount') : t('Fee'),
-    value: transactions[0].moduleAssetId === '2:0' ? asset.amount : fee,
-  };
-};
+import styles from './TransactionResultList.css';
+import { getTxConfig } from '../../utils';
 
 const Transactions = ({
   t, transactions, onSelectedRow, rowItemIndex, updateRowItemIndex, activeToken,
