+++ src/modules/search/components/TransactionResultList/index.js
@@ -1,0 +1,1 @@
+export { default } from './TransactionResultList';
