+++ src/modules/search/manager/searchBarManager.js
@@ -5,7 +5,7 @@ import { compose } from 'redux';
 import { withTranslation } from 'react-i18next';
 import { search } from '@common/utilities/api/search';
 import withData from '@common/utilities/withData';
-import SearchBar from './searchBar';
+import SearchBar from '../components/SearchBar/SearchBar';
 
 const mapStateToProps = state => ({
   activeToken: state.settings.token.active,
