+++ src/modules/search/utils/index.js
@@ -1,0 +1,2 @@
+export * from './searchUtils';
+export { default as getTxConfig } from './transactionListUtils';
